Fix organization invite + redirect
