+++ src/admin/users/core/actions/inviteListener.js
@@ -1,0 +1,58 @@
+import { fireStoreMainInstance } from '../../../../firebase'
+import { USER_INVITE_GET_SUCCESS } from '../usersActionTypes'
+import { redirectToProject } from '../../../project/utils/redirectToProject'
+import { addNotification } from '../../../notification/notifcationActions'
+import { redirectToOrganization } from '../../../organization/utils/redirectToOrganization'
+
+let stopListenForInvite
+export const listenForInvite = (inviteId, history) => (dispatch) => {
+    stopListenForInvite = fireStoreMainInstance
+        .collection('invites')
+        .doc(inviteId)
+        .onSnapshot(
+            (snapshot) => {
+                if (snapshot.exists) {
+                    const data = snapshot.data()
+                    dispatch({
+                        type: USER_INVITE_GET_SUCCESS,
+                        payload: data,
+                    })
+                    if (data.status === 'completed') {
+                        onInviteCompleted(data, history)
+                    }
+                } else {
+                    dispatch(
+                        addNotification({
+                            type: 'error',
+                            i18nkey: 'users.invitationGetFail',
+                        })
+                    )
+                }
+            },
+            (error) => {
+                // eslint-disable-next-line no-console
+                console.error(error)
+                dispatch(
+                    addNotification({
+                        type: 'error',
+                        i18nkey: 'users.invitationGetFail',
+                    })
+                )
+                return false
+            }
+        )
+}
+
+export const unsubscribeRealtimeInviteListener = () => () => {
+    stopListenForInvite && stopListenForInvite()
+}
+
+const onInviteCompleted = (invite, history) => {
+    if (invite.projectId) {
+        redirectToProject(null, invite.projectId, history)
+    } else if (invite.organizationId) {
+        redirectToOrganization(invite.organizationId, history)
+    } else {
+        throw new Error('Unmanaged invite type on completion')
+    }
+}
