+++ src/admin/users/core/actions/usersActions.js
@@ -1,27 +1,17 @@
-import {
-    fireStoreMainInstance,
-    nowTimestamp,
-    serverTimestamp,
-} from '../../../../firebase'
+import { fireStoreMainInstance } from '../../../../firebase'
 import {
     GET_USER_DETAILS_SUCCESS,
-    USER_INVITE_ADD,
-    USER_INVITE_GET_SUCCESS,
     USER_INVITE_REMOVE_SUCCESS,
     USER_INVITES_GET_SUCCESS,
     USERS_SET_FILTER,
 } from '../usersActionTypes'
-import { getPendingInvitesSelector, getUsersSelector } from '../usersSelectors'
+import { getUsersSelector } from '../usersSelectors'
 import {
     getMemberIds,
     getSelectedProjectIdSelector,
-    getSelectedProjectSelector,
 } from '../../../project/core/projectSelectors'
-import { getUserSelector } from '../../../auth/authSelectors'
-import { getDataFromProviderDataOrUser } from '../../../auth/authActions'
 import { addNotification } from '../../../notification/notifcationActions'
 import { editProject } from '../../../project/core/actions/editProject'
-import { redirectToProject } from '../../../project/utils/redirectToProject'
 
 export const getUserDetails = (uid) => (dispatch, getState) => {
     const usersDetails = getUsersSelector(getState())
@@ -78,49 +68,6 @@ export const setUsersFilter = (filterValue) => ({
     type: USERS_SET_FILTER,
     payload: filterValue.trim(),
 })
-
-let stopListenForInvite
-export const listenForInvite = (inviteId, history) => (dispatch) => {
-    stopListenForInvite = fireStoreMainInstance
-        .collection('invites')
-        .doc(inviteId)
-        .onSnapshot(
-            (snapshot) => {
-                if (snapshot.exists) {
-                    const data = snapshot.data()
-                    dispatch({
-                        type: USER_INVITE_GET_SUCCESS,
-                        payload: snapshot.data(),
-                    })
-                    if (data.status === 'completed') {
-                        redirectToProject(null, data.projectId, history)
-                    }
-                } else {
-                    dispatch(
-                        addNotification({
-                            type: 'error',
-                            i18nkey: 'users.invitationGetFail',
-                        })
-                    )
-                }
-            },
-            (error) => {
-                // eslint-disable-next-line no-console
-                console.error(error)
-                dispatch(
-                    addNotification({
-                        type: 'error',
-                        i18nkey: 'users.invitationGetFail',
-                    })
-                )
-                return false
-            }
-        )
-}
-
-export const unsubscribeRealtimeInviteListener = () => () => {
-    stopListenForInvite && stopListenForInvite()
-}
 
 export const getPendingInvites = () => (dispatch, getState) => {
     const projectId = getSelectedProjectIdSelector(getState())
